+++ src/main/java/racingcar/validation/exceptions/ValidationException.java
@@ -1,4 +1,4 @@
-package racingcar.validation;
+package racingcar.validation.exceptions;
 
 /**
  * packageName    : racingcar.validation
